requirements.txt
# Rush Medical College AI Admissions System
# Production Requirements

# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6

# Web Interface
streamlit==1.25.0

# Data Processing
openpyxl==3.1.2
xlsxwriter==3.1.2

# Essay Analysis
openai==1.3.0
PyPDF2==3.0.1
pdfplumber==0.9.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0
joblib==1.3.1

# Testing (optional)
pytest==7.4.0
pytest-cov==4.1.0